move proj to subdir for solution
