+++ RestaurantReviews.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.27130.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RestaurantReviews", "RestaurantReviews.csproj", "{08013ADC-4C49-4BD6-B6A2-0DB4129573C3}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RestaurantReviews", "RestaurantReviewsService\RestaurantReviews.csproj", "{79DFEE0B-A201-47BC-AB4F-CF202CAA99CD}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +11,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{08013ADC-4C49-4BD6-B6A2-0DB4129573C3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{08013ADC-4C49-4BD6-B6A2-0DB4129573C3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{08013ADC-4C49-4BD6-B6A2-0DB4129573C3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{08013ADC-4C49-4BD6-B6A2-0DB4129573C3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{79DFEE0B-A201-47BC-AB4F-CF202CAA99CD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{79DFEE0B-A201-47BC-AB4F-CF202CAA99CD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{79DFEE0B-A201-47BC-AB4F-CF202CAA99CD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{79DFEE0B-A201-47BC-AB4F-CF202CAA99CD}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
