+++ RestaurantReviewsService/appsettings.json
@@ -1,0 +1,18 @@
+﻿{
+  "Logging": {
+    "IncludeScopes": false,
+    "Debug": {
+      "LogLevel": {
+        "Default": "Warning"
+      }
+    },
+    "Console": {
+      "LogLevel": {
+        "Default": "Warning"
+      }
+    }
+  },
+  "JwtKey": "febb35af-7d8d-4b04-917a-4af4e59fc6b5",
+  "JwtIssuer": "http://yourdomain.com",
+  "JwtExpireMinutes": 30
+}
