--- Startup.cs
@@ -1,79 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.IdentityModel.Tokens;
-using System.Text;
-using Microsoft.EntityFrameworkCore;
-
-namespace RestaurantReviews
-{
-    public class Startup
-    {
-        public Startup(IConfiguration configuration)
-        {
-            Configuration = configuration;
-        }
-
-        public IConfiguration Configuration { get; }
-
-        // This method gets called by the runtime. Use this method to add services to the container.
-        public void ConfigureServices(IServiceCollection services)
-        {
-            // TODO
-            services.AddAuthentication(JwtBearerDefaults.AuthenticationScheme)
-                .AddJwtBearer("JwtBearer", options =>
-                {
-                    options.TokenValidationParameters = new TokenValidationParameters
-                    {                            
-                        ValidateIssuerSigningKey = true,
-                        IssuerSigningKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(Configuration["JwtKey"])),
-
-                        ValidateIssuer = true,
-                        ValidIssuer = Configuration["JwtIssuer"],
-
-                        ValidateAudience = true,
-                        ValidAudience = Configuration["JwtIssuer"],
-
-                        ValidateLifetime = true, //validate the expiration and not before values in the token
-
-                        ClockSkew = TimeSpan.FromMinutes(5) //5 minute tolerance for the expiration date
-                    };
-                });
-
-            services.AddDbContext<RestaurantReviews.Models.RestaurantReviewContext>(opt => opt.UseInMemoryDatabase("RestaurantReviews"));
-
-            services.AddMvc();
-        }
-
-        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
-        {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-            }
-
-            app.UseAuthentication();
-
-            app.UseStaticFiles();
-            
-            app.UseMvc();
-        }
-    }
-
-    public static class JwtSecurityKey
-    {
-        public static SymmetricSecurityKey Create(string secret)
-        {
-            return new SymmetricSecurityKey(Encoding.ASCII.GetBytes(secret));
-        }
-    }
-}
